% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_data.R
\name{formatting_data}
\alias{formatting_data}
\title{Remove/Replace Missing Slices}
\usage{
formatting_data(x, option = "remove")
}
\arguments{
\item{x}{}

\item{option}{}
}
\description{
Remove/Replace Missing Slices
}
