% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenz_plot.R
\name{frenz_plot}
\alias{frenz_plot}
\title{Quickly plotting a criterion network and the respondents accuracy.}
\usage{
frenz_plot(crit, accuracy_df,criterion="RLAS")
}
\arguments{
\item{crit}{criterion network returned after \link[Rfrenz:get_criterion]{get_criterion()} is run. the crit is used to plot a criterion network.}

\item{accuracy_df}{accuracy score dataframe that is return after \link[Rfrenz:sget_accuracy]{get_accuracy()}from Rfrenze is run. The accruacy_df is used to plot a boxplot summarising the data}

\item{criterion}{default="RLAS". the criterion name of the criterion being past in crit. is used as a titile for the graph.}
}
\description{
This function can be used to quickly plot a summary of the \link[Rfrenz:get_criterion]{get_criterion()}and \link[Rfrenz:get_accuracy]{get_accuracy()}functions in one place.
}
