% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenz_stats.R
\name{frenz_stats}
\alias{frenz_stats}
\title{Quickly get back basic descriptive statistics on Accuracy of Respondents.}
\usage{
frenz_stats(accuracy_df)
}
\arguments{
\item{accuracy_df}{The accuracy dataframe that results from the \link[Rfrenz:get_accuracy]{get_accuracy()} function.}
}
\description{
This function accepts a dataframe from the output of the \link[Rfrenz:get_accuracy]{get_accuracy()} function and is based on the Score coloumn. The function returns a dataframe with the basic descriptive statistics added to each observation. These statistics as as follows:
\describe{
\item{\strong{Mean}}{The Average Accuracy Score}
\item{\strong{Min}}{The Least Accurate Score}
\item{\strong{Max}}{The Most Accuracy Score}
\item{\strong{Median}}{The Median of the Accuracy Scores}
}
}
