% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_criterion.R
\name{get_criterion}
\alias{get_criterion}
\title{Get Criterion networks from all perceptions}
\usage{
get_criterion(sociomatrix, criterion)
}
\arguments{
\item{sociomatrix}{}

\item{criterion}{}
}
\description{
Get Criterion networks from all perceptions
}
\examples{

}
