% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hightech_advice-data.R
\docType{data}
\name{hightech_advice}
\alias{hightech_advice}
\title{High Tech Managers data on advice relations.}
\format{This data contains a list of socio-matrices. Each element in the list - a slice - contains a respondent's perception of the network. Each slice exists out of binary data where a '0' corresponds to no relationship obeservered and visa versa.}
\usage{
Rfrenz::hightech_advice
}
\description{
This data collected from the managers of a high-tech company.
The company manufactured high-tech equipment on the west coast of the United States and had just over 100 employees with 21 managers.
Each manager was asked to whom they go to for advice aswell as who they thought the other managers went to for advice. This was recorded and a dataset was made.
}
\examples{
data <- (Rfrenz::hightech_advice)
slice_6 <- data[[6]]
slice_6
}
\references{
Krackhardt D. (1987). Cognitive social structures. Social Networks, 9, 104-134.
}
\keyword{datasets}
