% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfrenz_acc.R
\name{rfrenz_acc}
\alias{rfrenz_acc}
\title{Get accuracy from slices}
\usage{
rfrenz_acc(dat, criterion = "RLAS", accuracy = "pearson",
  option = "remove")
}
\arguments{
\item{criterion}{A true (criterion) netwrok of the socio matricies to be computed.}

\item{accuracy}{An accuracy measure to be computed}

\item{option}{An option that determines how missing values will be dealt with. options: "zero", "remove", "replace"}

\item{slice}{A list of socio matrices that represent respondents.}
}
\value{
Returns a list of 2 items. The first is the criterion network. The second is a dataframe containing an accuracy score for each respondent
}
\description{
This function generates a criterion (true) network given all the perceptions perceived in a  network. The criterion network is then compared to all perceptions in order to determine th similarity between a respondent and their perception of the network.
}
\details{
Criterion (also called true or actual) networks are representative of the notion of shared or common perceptions of social structures from all observers within a network. Various different algorithms exist to derive these actual networks namely: Local Aggregate Structures (LAS), Consensus Structures and Expert Structures \insertCite{Krackhardt1987a}{Rfrenz}\insertCite{Cornelissen2019}{Rfrenz}.

\strong{Local Aggregate Structures (LAS)}

LAS is suitable in a context to generate which suggests that only the individuals directly involved in the relation determine the existence of the relation.
\describe{
\item{RLAS: Row-Dominated LAS}{An RLAS criterion is created by determining whether or not the perceiver observes the relationship to exist, even though the receiver does not always necessarily deem it to exist \insertCite{Krackhardt1987a}{Rfrenz}. Thus, the sender of the relationship will be assumed as correct. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{CLAS: Column-Dominated LAS.}{CLAS is created by determining whether or not the receiver of the relation perceives the relation to exist. Thus, the receiver of the relationship will be assumed as correct. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{ILAS: Intersection LAS}{ILAS criterion networks are created by determining whether or not both respondents perceive the tie to exist. Thus, both respondents have to agree on the relationship in order for it to exist \insertCite{Krackhardt1987a}{Rfrenz}. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{ULAS: Union LAS}{ULAS defines relations to exist if at least one of the respondents involved in the relation perceives it to exist \insertCite{Krackhardt1987a}{Rfrenz}. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
}

\strong{Consensus Structures}
\describe{
\item{GA: Global Aggregate}{Using Global Aggregate to generate actual networks is not so much focused on who perceived a specific relations, but rather how many perceived that relation. This is done by taking the sum of relationships that exists for each observer.}
\item{GAT: Global Aggregate with a threshold}{The threshold is set to measure the proportion of members that perceive a relation between the sender and receiver of the relationship. This corresponds to a “median response” notion of creating an actual network. Function used: \link[sna:consensus]{consensus}from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{GAV: Global Aggregate (valued)}{To gain a clearer understanding of the global perception of a relation, one could make use of proportions rather than showing only the presence or absence of a perceived relation (i.e. the GAT method). Thus, this method shows which percentage of observers perceived the relationship.}
}

\strong{Expert Structures}
\describe{
\item{SR: Single Reweight}{The reweight method takes consensus structure to create an initial criterion network. The reweight method is then applied to give a more accurate representation of the perceptions in the network. This is done is by performing consensus structure reduction on the data again but with different perceivers weighted according to their accuracy in perceiving relations in original criterion network. The single reweight method is when there is only one additional reduction done using the reweighted nodes. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{IR: Iterative Reweight}{Similar to Single Reweight except this method applies multiple reductions to weigh perceivers differently in each iteration. This can be useful when there is a disconnect between different perceptions which results in relations which are not accurate or don't exist. IR tries to find the most accurate perceivers in the network in order to give their perceptions of relations more validity when determining the criterion network. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{RB: Romney Batchelder}{Similar to Single Reweight method but with an exception to the motivation. Romney Batchelder theory's motivation lies in the measurement of shared beliefs in a group.  The criterion is built on the “cultural competency” weighted perceptions. Measuring “cultural competency” of respondents is done by determining the correlation between the responses of each individual and the aggregated responses of the group.}
\item{PCA: Principal Component Analysis.}{The PCA method for generating actual networks uses the first component of a network. PCA extracts shared themes among all the respondents within a network \insertCite{Butts2016}{Rfrenz}. PCA reduces the dimensions and relationships considered a data set while still maintaining the most important information within it. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{BAY: Bayesian}{Network accuracy model method can be used to generate the criterion network given the probability of false positives and false negatives. For more information see Bayesian. Function used: \link[sna:bbnam]{consensus}  from the SNA package \insertCite{Butts2016}{Rfrenz}.}
}
\strong{Accuracy Measures}
\describe{
 \item{Pearson Correlation}{The Pearson correlation test is done using a function from the SNA package called \link[sna:gcor]{gcor}. The SNA package states that “gcor” function is used to find the product-moment correlation between adjacency matrices \insertCite{Butts2016}{Rfrenz}. The “gcor” is a specialised form of a standard “cor” correlation. The ”gcor” function focuses more on graph data. Scores larger than 0 show a positive correlation whereas scores below 0 show a negative correlation.}
 \item{Spearman Correlation}{The Spearman correlation test is performed using the \link[stats:cor.test]{cor.test} function from the stats package. The estimate value is extracted from the results of the test. This value represents the similarity score computed from the correlation. According to the Stats package the rho statistic is used to compute the estimate value of the ranked based measure of association \insertCite{DeNooy2016}{Rfrenz}. A Spearman correlation test is best suited for data that can be ranked. Scores larger than 0 show a positive correlation whereas scores below 0 show a negative correlation.}
 \item{Jaccard Correlation}{The Jaccard similarity coefficient is performed using the \link[jaccard:jaccard]{jaccard} function that is found in the Jaccard package. This function will result in a number between 1 and 0. The closer the value is to 1 the higher the similarity is between the two matrices. The Jaccard similarity coefficient is used in social science as it shows the occurrences of agreements between binary sets of data \insertCite{Marineau2016b}{Rfrenz}.}
 \item{Kendall Rank Correlation}{The Kendall rank correlation is performed using the \link[Kendall:Kendall]{kendall} function that is found in the Kendall package.  The Kendall rank correlation or more commonly known as Kendall's tau, is a correlation tool used to measure the ordinal association between two measured variables \insertCite{Kendall1938}{Rfrenz}. The “Kendall” function computes a p-value and if it determined to be significant such that variables are dependent, and therefore there are ties. Then it will checks the similarity between these variables. The Kendall Tau statistic is extracted from the results of this function. }
 \item{S14 Similarity Index}{The s14 similarity index is performed using the \link[cssTools:s14]{s14} function that is found in the cssTools package. The s14 function computes the similarity between the two square matrices.  S14 is useful for network data as it shows appropriate sensitivity to small changes and does not distort at extremes \insertCite{Gower1986}{Rfrenz}.}
 \item{MRQAP with Double-Semi-Partialing  (DSP)}{MRQAP with DSP is performed using the \link[asnipe:mrqap.dsp]{mrqap.dsp} function that is found in the Asnipe package. This function computes the regression coefficient using the DSP method. The method randomises the p-value with 1000 permutation in order to determine if the matrices are dependent or independent \insertCite{Dekker2003}{Rfrenz}.}
 \item{Structural Correlation (gscor)}{The structural correlation between two or more graphs can be determined using the \link[sna:gscor]{gscor} function found in the SNA package. The gscor function finds the product-movement structural correlation. This indicates that it looks at the structure of each network and determines how similar the structure is as opposed to if specific ties were similar between nodes \insertCite{Butts2001}{Rfrenz}. The node labels in the network are permutated in order to determine how similar the structure is to the true network \insertCite{Butts2001}{Rfrenz}.}
 \item{Cohens Kappa}{The Cohens Kappa coefficient can be determined using the \link[psych:cohen.kappa]{cohen.kappa} function that can be found in the psych package. This accuracy measure takes the possibility of chance into account (i.e., someone could have merely guessed about a relationship within a given network and the measure takes that probability into account) \insertCite{Cohen1960}{Rfrenz}. Target accuracy measures refer to how accurate a perceiver's agreement would be towards a certain target, where the target would be the relationship between i and j \insertCite{Neal2016}{Rfrenz}.}
 \item{Local Accuracy}{The local accuracy is computed by comparing a person's perceptions to a Row dominated local aggregate structure (RLAS). local accuracy tries to see how accurate a person is at knowing the perceptions in a social network based on how they think the people in that network perceive them \insertCite{Casciaro1999}{Rfrenz}. A pearson correlation is performed between a row of a person's perceptions and a row of the RLAS criterion.}
 \item{Dyadic Accuracy}{Dyadic accuracy looks at how accurately an individual can perceive a possible dyadic relation between two individuals. To do this it takes one person perceptions and measures its perception of dyadic relations according to the ``true'' network \insertCite{Bondonio1998}{Rfrenz}.}
 \item{Triadic (Pearson, Spearman, Distance)}{The triadic accuracy compares the triadic structure of the criterion network and the respondent by means of either a Pearson correlation, Spearman's rank correlation, or Euclidean distance measure. This approach was proposed by \insertCite{Frenz2019;textual}{Rfrenz}  as a means to compare the degree of similarity between the triad census of the actual social network and the triad census of a respondents cognitive slice}
}
}
\examples{
r1 <- matrix(c(0,1,0,1,0,
0,0,1,1,0,
1,0,0,1,1,
0,1,0,0,1,
0,1,0,1,0),nrow=5, ncol=5)
r2 <- matrix(c(0,1,1,1,0,
            0,0,1,1,0,
            1,0,0,1,0,
            1,1,0,0,1,
            1,1,0,1,0),nrow=5, ncol=5)
r3 <- matrix(c(0,1,1,1,0,
            1,0,1,1,0,
            1,1,0,0,0,
            0,1,0,0,1,
            1,1,0,1,0),nrow=5, ncol=5)
r4 <- matrix(c(0,1,0,1,0,
            1,0,1,1,1,
            1,0,0,1,1,
            0,1,0,0,0,
            0,0,0,1,0),nrow=5, ncol=5)
r5 <- matrix(c(NA,NA,NA,NA,NA,
            NA,NA,NA,NA,NA,
            NA,NA,NA,NA,NA,
            NA,NA,NA,NA,NA,
            NA,NA,NA,NA,NA),nrow=5, ncol=5)

list_respondent <- list(r1,r2,r3,r4,r5)

rfrenz_acc<- rfrenz_acc(list_respondent, acc = "pearson", criterion_type = "RLAS")

}
