% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_data.R
\name{formatting_data}
\alias{formatting_data}
\title{Remove/replace missing slices}
\usage{
formatting_data(dat, option = "remove")
}
\arguments{
\item{dat}{A list of socio-matrices containing all of the perceptions of a network}

\item{option}{Option whether to remove or replace missing values ("remove", "replace", "zero")}
}
\value{
A list of socio-matracies with the removed/replaced slices within the list
}
\description{
This function removes any missing perceptions (slices) from a list of socio-matrices. Three different options exist:
removing missing slices, replacing missing slices with central graph or replacing missing values with 0.
}
\details{
Missing data can affect the formation of the actual networks. In order to deal with this,
missing perceptions can be set to a blank perception (0),
or replaced with an estimate value, or the missing respondents can be omitted.
\describe{
\item{"remove"}{This option will omit the missing values completely. Omitting the missing respondents results in effectively sampling the original data \insertCite{Borgatti2006}{Rfrenz}.}
\item{"replace"}{This option will replace the missing values of the obeserver with the central graph consensus of the network.}
\item{"zero"}{This option will replace all the missing values of the obeserver with the value 0. This however, can result in inaccurate outcomes \insertCite{Krackhardt2002}{Rfrenz}.}
}
}
\examples{
# Creating list of matrices for exmaple
r1 <- matrix(c(0,1,0,1,0,
0,0,1,1,0,
1,0,0,1,1,
0,1,0,0,1,
0,1,0,1,0),nrow=5, ncol=5)
r2 <- matrix(c(0,1,1,1,0,
               0,0,1,1,0,
               1,0,0,1,0,
               1,1,0,0,1,
               1,1,0,1,0),nrow=5, ncol=5)
r3 <- matrix(c(0,1,1,1,0,
               1,0,1,1,0,
               1,1,0,0,0,
               0,1,0,0,1,
               1,1,0,1,0),nrow=5, ncol=5)
r4 <- matrix(c(0,1,0,1,0,
               1,0,1,1,1,
               1,0,0,1,1,
               0,1,0,0,0,
               0,0,0,1,0),nrow=5, ncol=5)
r5 <- matrix(c(NA,NA,NA,NA,NA,
               NA,NA,NA,NA,NA,
               NA,NA,NA,NA,NA,
               NA,NA,NA,NA,NA,
               NA,NA,NA,NA,NA),nrow=5, ncol=5)

list_respondent <- list(r1,r2,r3,r4,r5)    
         
formatted_data <- formatting_data(list_respondent, option = "replace")

}
\references{
\insertAllCited{}
}
\author{
Christiaan van Rensburg, Alex Cawood, Marette Theron
}
