% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frenz_plot.R
\name{frenz_plot}
\alias{frenz_plot}
\title{Frenz Plot}
\usage{
frenz_plot <- function(crit, accuracy_df,criterion="RLAS")
}
\arguments{
\item{crit}{criterion network returned after get_criterion() is run. the crit is used to plot a criterion network.}

\item{accuracy_df}{accuracy score dataframe that is return after get_accruacy() from Rfrenze is run. the accruacy_df is used to plot a boxplot summarising the data}

\item{criterion}{default="RLAS". the criterion name of the criterion being past in crit. is used as a titile for the graph.}
}
\description{
function used to quickly plot a summary of the get_crit() and get_accuracy() functions in one place.
}
