% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accuracy.R
\name{get_accuracy}
\alias{get_accuracy}
\title{Get Accuracy of Respondents}
\usage{
get_accuracy(slice, criterion, acc = "pearson", criterion_type = "")
}
\arguments{
\item{criterion_type}{}
}
\value{
A dataframe of respondents and there accuracy scores
}
\description{
Get Accuracy of Respondents
}
