% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accuracy.R
\name{get_accuracy}
\alias{get_accuracy}
\title{Get Accuracy of Respondents}
\usage{
get_accuracy(dat, criterion, acc = "pearson", criterion_type = NA)
}
\arguments{
\item{dat}{A list of socio-matrices that represent respondents.}

\item{criterion}{A socio-matrix that represents the criterion network (i.e., a network that best represents the relations in a network) that is computed from all respondents.}

\item{acc}{The name of the accuracy measure to be computed. Options: “pearson”, “spearman”, “jaccard”, “kendall”, “s14”, “mrqap”, “gscor”, “CohenK”, “local”, “triadic pearson”, “triadic spearman”, “triadic distance”.}

\item{criterion_type}{The name of the criterion that has been used.}
}
\value{
A dataframe of respondents and there accuracy scores.
}
\description{
The purpose of this function is to get the accuracy scores for each respondent in a social network based on their perception of the network. This is done by checking the similarity between their perception and a criterion network (i.e., a network that best represents the relations in a network).
}
\details{
The details of each accuracy measure will be explored in this section.
\describe{
 \item{Pearson Correlation}{The Pearson correlation test is done using a function from the SNA package called \link[sna:gcor]{gcor}. The SNA package states that “gcor” function is used to find the product-moment correlation between adjacency matrices \insertCite{Butts2016}{Rfrenz}. The “gcor” is a specialised form of a standard “cor” correlation. The ”gcor” function focuses more on graph data. Scores larger than 0 show a positive correlation whereas scores below 0 show a negative correlation.}
 \item{Spearman Correlation}{The Spearman correlation test is performed using the \link[stats:cor.test]{cor.test} function from the stats package. The estimate value is extracted from the results of the test. This value represents the similarity score computed from the correlation. According to the Stats package the rho statistic is used to compute the estimate value of the ranked based measure of association \insertCite{DeNooy2016}{Rfrenz}. A Spearman correlation test is best suited for data that can be ranked. Scores larger than 0 show a positive correlation whereas scores below 0 show a negative correlation.}
 \item{Jaccard Correlation}{The Jaccard similarity coefficient is performed using the \link[jaccard:jaccard]{jaccard} function that is found in the Jaccard package. This function will result in a number between 1 and 0. The closer the value is to 1 the higher the similarity is between the two matrices. The Jaccard similarity coefficient is used in social science as it shows the occurrences of agreements between binary sets of data \insertCite{Marineau2016b}{Rfrenz}.}
 \item{Kendall Rank Correlation}{The Kendall rank correlation is performed using the \link[Kendall:Kendall]{kendall} function that is found in the Kendall package.  The Kendall rank correlation or more commonly known as Kendall's tau, is a correlation tool used to measure the ordinal association between two measured variables \insertCite{Kendall1938}{Rfrenz}. The “Kendall” function computes a p-value and if it determined to be significant such that variables are dependent, and therefore there are ties. Then it will checks the similarity between these variables. The Kendall Tau statistic is extracted from the results of this function. }
 \item{S14 Similarity Index}{The s14 similarity index is performed using the \link[cssTools:s14]{s14} function that is found in the cssTools package. The s14 function computes the similarity between the two square matrices.  S14 is useful for network data as it shows appropriate sensitivity to small changes and does not distort at extremes \insertCite{Gower1986}{Rfrenz}.}
 \item{MRQAP with Double-Semi-Partialing  (DSP)}{MRQAP with DSP is performed using the \link[asnipe:mrqap.dsp]{mrqap.dsp} function that is found in the Asnipe package. This function computes the regression coefficient using the DSP method. The method randomises the p-value with 1000 permutation in order to determine if the matrices are dependent or independent \insertCite{Dekker2003}{Rfrenz}.}
 \item{Structural Correlation (gscor)}{The structural correlation between two or more graphs can be determined using the \link[sna:gscor]{gscor} function found in the SNA package. The gscor function finds the product-movement structural correlation. This indicates that it looks at the structure of each network and determines how similar the structure is as opposed to if specific ties were similar between nodes \insertCite{Butts2001}{Rfrenz}. The node labels in the network are permutated in order to determine how similar the structure is to the true network \insertCite{Butts2001}{Rfrenz}.}
 \item{Cohens Kappa}{The Cohens Kappa coefficient can be determined using the \link[psych:cohen.kappa]{cohen.kappa} function that can be found in the psych package. This accuracy measure takes the possibility of chance into account (i.e., someone could have merely guessed about a relationship within a given network and the measure takes that probability into account) \insertCite{Cohen1960}{Rfrenz}. Target accuracy measures refer to how accurate a perceiver's agreement would be towards a certain target, where the target would be the relationship between i and j \insertCite{Neal2016}{Rfrenz}.}
 \item{Local Accuracy}{The local accuracy is computed by comparing a person's perceptions to a Row dominated local aggregate structure (RLAS). local accuracy tries to see how accurate a person is at knowing the perceptions in a social network based on how they think the people in that network perceive them \insertCite{Casciaro1999}{Rfrenz}. A pearson correlation is performed between a row of a person's perceptions and a row of the RLAS criterion.}
 \item{Dyadic Accuracy}{Dyadic accuracy looks at how accurately an individual can perceive a possible dyadic relation between two individuals. To do this it takes one person perceptions and measures its perception of dyadic relations according to the ``true'' network \insertCite{Bondonio1998}{Rfrenz}.}
 \item{Triadic (Pearson, Spearman, Distance)}{The triadic accuracy compares the triadic structure of the criterion network and the respondent by means of either a Pearson correlation, Spearman's rank correlation, or Euclidean distance measure. This approach was proposed by \insertCite{Frenz2019;textual}{Rfrenz}  as a means to compare the degree of similarity between the triad census of the actual social network and the triad census of a respondents cognitive slice}
}
}
\examples{

}
\references{
\insertAllCited{}
}
