% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accuracy.R
\name{get_accuracy}
\alias{get_accuracy}
\title{Get Accuracy of Respondents}
\usage{
get_accuracy(slice, criterion, acc = "pearson", criterion_type = "")
}
\arguments{
\item{slice}{A list of socio matricies that represent respondents.}

\item{criterion}{A socio matrix that represnts the true network computed from all respondents}

\item{acc}{The accuracy measure to be computed}

\item{criterion_type}{The criterion type that has been used}
}
\value{
A dataframe of respondents and there accuracy scores
}
\description{
Get Accuracy of Respondents
}
