% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_criterion.R
\name{get_criterion}
\alias{get_criterion}
\title{Get Criterion networks from all perceptions}
\usage{
get_criterion(x, criterion)
}
\arguments{
\item{criterion}{}

\item{sociomatrix}{}
}
\description{
Get Criterion networks from all perceptions
}
\examples{

}
