% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_criterion.R
\name{get_criterion}
\alias{get_criterion}
\title{Get Criterion networks from all perceptions}
\usage{
get_criterion(x, criterion=”RLAS”)
}
\arguments{
\item{x}{A socio-matrix containing all of the perceptions of a network}

\item{criterion}{The criterion network needed to be generated (“RLAS”, “CLAS”, “ILAS”, “ULAS”, “GA”, “GAT”, “GAV”, “SR”, “IR”, “PCA”, “RB”, “BAY”)}
}
\value{
A socio-matrix representing the Criterion ('true') Network structure
}
\description{
This function generates a criterion (true) network given all the perceptions perceived in a  network. Several algorithms for determining criterion networks are available to best suit the context of the network.
}
\details{
Criterion (also called true or actual) networks are representative of the notion of shared or common perceptions of social structures from all observers within a network. Various different algorithms exist to derive these actual networks namely: Local Aggregate Structures (LAS), Consensus Structures and Expert Structures \insertCite{Krackhardt1987a}{Rfrenz}\insertCite{Cornelissen2019}{Rfrenz}.

\strong{Local Aggregate Structures (LAS)}

LAS is suitable in a context to generate which suggests that only the individuals directly involved in the relation determine the existence of the relation.
\describe{
\item{RLAS: Row-Dominated LAS}{An RLAS criterion is created by determining whether or not the perceiver observes the relationship to exist, even though the receiver does not always necessarily deem it to exist \insertCite{Krackhardt1987a}{Rfrenz}. Thus, the sender of the relationship will be assumed as correct. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{CLAS: Column-Dominated LAS.}{CLAS is created by determining whether or not the receiver of the relation perceives the relation to exist. Thus, the receiver of the relationship will be assumed as correct. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{ILAS: Intersection LAS}{ILAS criterion networks are created by determining whether or not both respondents perceive the tie to exist. Thus, both respondents have to agree on the relationship in order for it to exist \insertCite{Krackhardt1987a}{Rfrenz}. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{ULAS: Union LAS}{ULAS defines relations to exist if at least one of the respondents involved in the relation perceives it to exist \insertCite{Krackhardt1987a}{Rfrenz}. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
}

\strong{Consensus Structures}
\describe{
\item{GA: Global Aggregate}{Using Global Aggregate to generate actual networks is not so much focused on who perceived a specific relations, but rather how many perceived that relation. This is done by taking the sum of relationships that exists for each observer.}
\item{GAT: Global Aggregate with a threshold}{The threshold is set to measure the proportion of members that perceive a relation between the sender and receiver of the relationship. This corresponds to a “median response” notion of creating an actual network. Function used: \link[sna:consensus]{consensus}from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{GAV: Global Aggregate (valued)}{To gain a clearer understanding of the global perception of a relation, one could make use of proportions rather than showing only the presence or absence of a perceived relation (i.e. the GAT method). Thus, this method shows which percentage of observers perceived the relationship.}
}

\strong{Expert Structures}
\describe{
\item{SR: Single Reweight}{The reweight method takes consensus structure to create an initial criterion network. The reweight method is then applied to give a more accurate representation of the perceptions in the network. This is done is by performing consensus structure reduction on the data again but with different perceivers weighted according to their accuracy in perceiving relations in original criterion network. The single reweight method is when there is only one additional reduction done using the reweighted nodes. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{IR: Iterative Reweight}{Similar to Single Reweight except this method applies multiple reductions to weigh perceivers differently in each iteration. This can be useful when there is a disconnect between different perceptions which results in relations which are not accurate or don't exist. IR tries to find the most accurate perceivers in the network in order to give their perceptions of relations more validity when determining the criterion network. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{RB: Romney Batchelder}{Similar to Single Reweight method but with an exception to the motivation. Romney Batchelder theory's motivation lies in the measurement of shared beliefs in a group.  The criterion is built on the “cultural competency” weighted perceptions. Measuring “cultural competency” of respondents is done by determining the correlation between the responses of each individual and the aggregated responses of the group.}
\item{PCA: Principal Component Analysis.}{The PCA method for generating actual networks uses the first component of a network. PCA extracts shared themes among all the respondents within a network \insertCite{Butts2016}{Rfrenz}. PCA reduces the dimensions and relationships considered a data set while still maintaining the most important information within it. Function used: \link[sna:consensus]{consensus} from the SNA package \insertCite{Butts2016}{Rfrenz}.}
\item{BAY: Bayesian}{Network accuracy model method can be used to generate the criterion network given the probability of false positives and false negatives. For more information see Bayesian. Function used: \link[sna:bbnam]{consensus}  from the SNA package \insertCite{Butts2016}{Rfrenz}.}
}
}
\references{
\insertAllCited{}
}
\author{
Christiaan van Rensburg, Alex Cawood, Marette Theron
}
