% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(path, save = FALSE, path_out)
}
\arguments{
\item{path}{Path to the directory where txt files are stored.}

\item{save}{Default False. TRUE/FALSE  The option to save your matrix into an RDS. If set to TRUE you need to pass a path to "path_out" where your RDS will be stored.}

\item{path_out}{The Path to a driectory where you want to store the RDS.}
}
\description{
Get Data
}
